pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prismata"
version = "0.2.0"
description = "Prismata: Multi-domain video event detection framework using temporal deep learning"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Vijay Ponduru", email = "ponduru@gmail.com"}
]
keywords = ["prismata", "event-detection", "sports", "video-analysis", "deep-learning", "pytorch", "surveillance"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "opencv-python>=4.8.0",
    "ffmpeg-python>=0.2.0",
    "albumentations>=1.3.0",
    "numpy>=1.24.0",
    "pillow>=9.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.0.270",
    "mypy>=1.0.0",
]
labeling = [
    "streamlit>=1.25.0",
    "yt-dlp>=2023.0.0",
]
tracking = [
    "wandb>=0.15.0",
]
all = [
    "prismata[dev,labeling,tracking]",
]

[project.scripts]
prismata-train = "scripts.train:main"
prismata-detect = "scripts.detect:main"
prismata-label = "scripts.label:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src"
